Add EnvPrefix()
Same functionality as Env() except a prefix is used when looking up
environment variables to replace with. So a key that is looked up with
Env() as FOO_BAR will be BAZ_FOO_BAR if EnvPrefix("baz") is used.

Also updated TestEnv to ensure that the previously set value is
overwritten by the environment variable.